+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [unreleased]
+## [1.1.0] - 2023-11-19
 
 ### Added
 
@@ -53,7 +53,7 @@
 _🍰 Initial release_
 
 
-[unreleased]: https://github.com/phoenixr-codes/mcnbt/compare/v1.0.0...HEAD
+[1.1.0]: https://github.com/phoenixr-codes/mcnbt/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/phoenixr-codes/mcnbt/compare/v0.2.0...v1.0.0
 [0.2.0]: https://github.com/phoenixr-codes/mcnbt/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/phoenixr-codes/mcnbt/releases/v0.1.0